feat: remove Response iterator, make content always sequence
The variable return type of content (Content | Sequence[Content]) was inconsistent with all the other accessors, and it prevented using response.content as an iterator. If response.content is always a sequence, then there's no need for response to be an iterator - you can just do `for item in response.content` instead.