feat: add unified examples.mdx for LLM consumption
Adds a single mdx document (6k tokens) with a broad ranging set of
examples all concatenated together. The idea is that this is very
helpful as a unified content blog that can be passed to LLMs to prime
them on Mirascope's API design and philosophy (as demonstrated in
practice), so that said LLMs can then be better rubber ducks for
exploring interface changes or improvements.

Works pretty well (Claude gives helpful / plausible suggestions on how
to e.g. integrate MCP with Mirascope v2 given just this context for
priming)