feat: implement StreamResponse.chunk_stream() with testing
Implements the core logic of llm.StreamResponse, with extensive testing,
and for both sync and async cases.

The tests are parameterized over sync and async cases, so all tests hit
both code paths. There are helper functions that make this easy to
manage.

Currently text and thinking support are both added, but not tool calls,
as we haven't established patterns for deserializing the call args and I
didn't feel this was the place to start.

One thing I noticed while implementing is that we don't really have a
good way for the StreamResponse to get the finish reason, so I will
address that in a followon.