feat: add google streaming
Also: Change in behavior for google, stop removing trailing newlines for
consistency, let Google's behavior be different (always adds a newline
to text)

Rationale: I put it in for call, but it is annoying to add to streaming
(would need to buffer chunks and lag behind the api by 1 chunk to make
sure we only strip newline off the last part, not parts in the middle of
the content).

Ultimately while I like the idea of enforcing model consistency, I'm not
sure it's worth it, esp as I haven't tested this across other Google
models to know if it's a general property or just a quirk of Gemini 2.0.

Open to revisiting it.

Also; factored out some shared helpers for preparing google responses.