feat: add ToolSchema / ToolParameterSchema with tests
ToolSchema is a Pydantic/JSONSchema based approach for representing tool
defs. The implementation is based on v1's approach. There is a class
method on the schema for constructing one from function signature, and a
good number of tests.

There is overlap with BaseTool's fields. My plan in the next PR is to
add schema onto BaseTool, and have the BaseTool properties become
accessors that pluck from the schema, so we can keep the BaseTool API
but also the nice centralization on the ToolSchema model that I like.