refactor: store tool call args as string in content
intention: ensure that processing a LLM response with tool calls can't
cause a validation error (even if llm provided invalid json). instead,
the validation error may occur when calling the tool, or potentially
(depending on provider) when attempting to include the malformed tool
call in a subsequent messsage to the llm.