refactor: add union overloads to client/llm apis
Rationale: This makes it actually type safe to write code like
`client.call(**kwargs)` where kwargs may include format.

We were already using this pattern in our test code, however the type
error was obscured because we were getting the kwargs typed as `Any` via
`request.getfixturevalue()`. Since the pattern is actually useful, we
should support it in a type-safe way. As an upside, we can also clean up
the LLM implementation.