feat: e2e tool usage (including context tools) with tests
This is one part adding the end to end tests, one part finishing
fleshing out NotImplemented parts of the interface (particularly where
context tools are involved). Added unit tests where necessary to get
full coverage (mixed context/non context tools case) and to fully pin
down the intended contract (ensuring that ctx arg is filtered out of the
ToolSchema). Getting the tool schemas to be identical regardless of the
ctx arg is necessary, otherwise the context calls won't match the
cassettes.

Aside: As I continue working on this, it gets clearer and clearer that
we should unify `@llm.tool` and `@llm.context_tool` (and others)
assuming that we can. It's so easy to mix up and add the ctx argument
but forget to switch decorators.