refactor: unify params
Makes the Mirascope interface more consistent across providers,
removes thorny type issues and the need for methods like
`assumed_safe_llm_create`, ensures that we can both store params on
responses and have cross-provider response resume functionality.

When a param is not supported for a given provider (e.g. seed for
anthropic), I suggest we generally log a warning in preference to
raising llm.FeatureNotSupportedError, though we can cross that bridge
when we (shortly) come to it.