refactor: pass toolkits not tools
- Client APIs can now take either toolkit or tools
- Response constructor **must** take toolkit
- Call decorator still takes tools not toolkit

This is a non-breaking API change (calls to client can still just pass
tools). It does get a nice performance improvement because we now only
construct toolkits in two cases:
- When call decorator is invoked and creates a Call
- When user has called client with tools rather than toolkit

All other cases, we pass the toolkit by reference rather than
constructing it anew. This is good because constructing toolkit is
O(number of tools). We should still probably put a LRU cache in, but
it's much less urgent now that the library code constructs toolkits only
where truly necessary.