feat: unbreak google when penalty params are present
In testing, the 2.0 series accepts presence parameters, and the 2.5
series fails when they are used. So, I updated the implementation to log
a warning and skip parameter when they can't be used.

Once the next generation comes out, we can infer whether Google is
dropping support for this long term or if 2.5 is the exception, and
maybe switch to "default exclude" for the parameter.

For now I did a coverage exclusion for convenience, because the e2e
testing setup doesn't make it easy to run the same tests for multiple
models with the same provider (would need to change how we name snapshot
files and cassettes to disambiguate). However I'm open to figuring
something else out.