refactor: improve type safety of google utils
Now we use the actual `genai_types.GenerateContentConfigDict` type when
constructing the google config, meaning that if we typo a property
assignment (e.g. trying to pass `max_tokens` instead of
`max_output_tokens`), then we get a type error.

The old code constructed the dict without type safety, and then created
the pydantic model version (GenerateContentConfig). If we want to, we
can also construct the pydantic model for extra safety, but I think it's
in keeping with our design approach to depend on the type system here.