feat: stub out openai:responses provider
We add a new provider, "openai:responses" with a corresponding
`OpenAIResponsesClient`. Currently the client raises NotImplementedError
for every method, so this is just showing how it will fit into the
codebase (for now).

Currently I'm leaving the chat completions client as just "openai"
provider, so this approach privileges chat completions as the default
option and naming reflects this (`OpenAIClient` vs
`OpenAIResponsesClient`). I don't see this as the final design, more
just a starting point while the responses client gets developed. Before
launching beta we should revisit this. Added a TODO to this effect.