nit: remove inconsistent prepare_request docstrings
Removes docstrings for the openai prepare_request methods, these are
internal-only utils in the _utils module, and the docstring was
basically just a redeclaration of the type signature anyway. This makes
it consistent with how we handle the other prepare methods for google
and anthropic. Also, rename the methods to distinguish which API they
serve.