fix: attempt to compensate for a google-specific api bug
When Google hits its max token limits during the pre-response thinking
phase, and is in streaming mode, it emits no events at all, so we
incorrectly set the finish reason to None. This is a strange behavior so
I have added a patch that sets finish reason to max tokens in this case.
I hoped I could do so without adding any other bugs, however there is
one false positive in the sync (not async!) streaming case where the
model is prompted to emit no output at all. I think this is an OK
tradeoff to make although I'm sad the fix isn't perfect. Behavior
documented in a provider-specific e2e test.