refactor: unify Image and ImageURL
I feel it is a lot cleaner if we have one Image content type and it has
a `source` field which specifies the source flexibly. I took inspiration
from Anthropic's ImageBlockParam which I feel is well-designed.

I removed Image.id as it is present for speculative purposes
(referencing generated image outputs), but image generation is not in
scope. Since it is an optional field, we can add it back in when we
actually have a plan for it.