refactor: Unify `Audio` and `AudioURL`
Similar reasoning and approach as previous commit. However, in this case
I went further and removed the `AudioURLSource` option. Why? Because the
providers that support audio (google and openai:completions) don't
support referencing audio by URL. OpenAI requires that it be bytes64,
and Google supports bytes64 or upload to their proprietary files API
(which is not a url).

So, no reason including AudioURL if none of the providers yet support
it. I structured the class so we can always add it in as an alternative
source later.

I also removed the transcript option, for similar reasons: none of the
providers take transcripts for uploaded audio, and we aren't looking
ahead to generated audio just yet. We can always add transcript in when
we are ready to use the field.