test: add test_resume_with_override_thinking_and_tools
Also fix up some implementation issues, particularly around checking
`if part.type == "thought" and encode_thoughts` which would cause the
thought to trigger a fallback exception if not `encode_thoughts`. Now
its changed to `if part.type == "thought":` with the `encode_thoughts`
check nested, which results in the correct behavior that if we are not
encoding_thoughts, we do nothing.