refactor: make RawStreamEventChunk and RawContentChunk clearer
This refactor makes RawContentChunk a clear peer to RawStreamEventChunk
(previoulsy RawChunk which was confusing) and drops it from
AssistantContentChunk since it doesn't actually reconstruct content. I
think treating the raw_content as a bit "under the hood" makes more
sense. We now don't include the RawContentChunk in chunk_stream which is
consistent with the RawStreamEventChunk.