test: ensure VCR overwrite error won't be snapshotted
During presubmit self review I noticed that a VCR overwrite error had made it into
snapshots, fixed it in the offending PR but adding this so they can't
sneak in again. Tested this locally by running --fix in a state where it
would otherwise put the VCR exception into the snapshot.