test: refactor all e2e tests to catch exceptions
Now all the e2e tests catch exceptions rather than failing if a provider
throws an error (excepting VCR over-write errors which the exception
serializer re-throws).

The intent of this change is it becomes possible to e.g. add a partially
implemented provider to the list of providers being tested, knowing that
it will throw exceptions on many tests, and then progressively
implementing those features. While still maintaining a green build every
step of the way.