fix: resolve import error with python 3.13.9
In python 3.13.9, attempting to import Mirascope results in the
following error:

> TypeError: can only concatenate list (not "tuple") to list

After extensive investigation, it turns out this is due to a bug in
cpython, discussed here: https://github.com/python/cpython/issues/138859

The issue involves paramspecs with `default=...` (as our `AnyP` uses)
when that paramspec is not the last type argument. It has already been
fixed, but the fix has not yet made it into a 3.13 release (likely will
be fixed in 3.13.10).

To ensure immediate compatibility with 3.13, we can work around the bug
by re-ordering the paramspec argument to the last generic type arg to
`ContextTool` and `AsyncContextTool`.

This should fully resolve issues with python 3.13. I've added 3.13 to
the list of python versions under test.

Props to chatgpt-pro for ingesting a (still quite complex) minimal repro
and finding the correct underlying issue in cpython. Claude hallucinated
"helpful" but incorrect solutions, and Gemini was stumped too.
chatgpt-pro took 9 minutes to do research, but came back with
information that was actually correct.