chore: fix unknown type issues in client code
Setting the base context_var property as ContextVar[Any] is not my first
choice (ContextVar[Self] is more accurate) but it was necessary to work
around some variance issues (the TypeVar for ContextVar is invariant).
Since it's a private property I think it's okay. (Also we are planning
to move the clients out of the public API anyway I think.)